# Chapter03/payments/internal/rest/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Payments API",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "PaymentsService"
    }
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/payments": {
      "post": {
        "summary": "Authorize a future payment",
        "operationId": "authorizePayment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentspbAuthorizePaymentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/paymentspbAuthorizePaymentRequest"
            }
          }
        ],
        "tags": [
          "Payment"
        ]
      }
    },
    "/api/payments/invoices/{id}/pay": {
      "put": {
        "summary": "Pay an invoice",
        "operationId": "payInvoice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentspbPayInvoiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PaymentsServicePayInvoiceBody"
            }
          }
        ],
        "tags": [
          "Invoice"
        ]
      }
    }
  },
  "definitions": {
    "PaymentsServicePayInvoiceBody": {
      "type": "object"
    },
    "paymentspbAdjustInvoiceResponse": {
      "type": "object"
    },
    "paymentspbAuthorizePaymentRequest": {
      "type": "object",
      "properties": {
        "customerId": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "paymentspbAuthorizePaymentResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "paymentspbCancelInvoiceResponse": {
      "type": "object"
    },
    "paymentspbConfirmPaymentResponse": {
      "type": "object"
    },
    "paymentspbCreateInvoiceResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "paymentspbPayInvoiceResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}